Extract callback parsers and UI helpers into bot/handlers package
Move parse_download_callback, parse_summary_option into
bot/handlers/callback_parsing and format_bytes, format_eta,
safe_edit_message, send_long_message into bot/handlers/common_ui.

telegram_callbacks imports from the new modules and drops ~160 lines
of inline definitions while preserving the public interface.
