+++ bot/handlers/__init__.py
@@ -1,0 +1,1 @@
+"""Feature-oriented Telegram handler helpers."""
