+++ bot/handlers/common_ui.py
@@ -1,0 +1,93 @@
+"""Shared Telegram UI helpers for callback/command flows."""
+
+from __future__ import annotations
+
+import logging
+
+from telegram.error import BadRequest, NetworkError, TimedOut
+
+
+def format_bytes(bytes_value):
+    """Formats bytes to human readable string."""
+    if bytes_value is None:
+        return "?"
+    for unit in ["B", "KB", "MB", "GB"]:
+        if bytes_value < 1024:
+            return f"{bytes_value:.1f} {unit}"
+        bytes_value /= 1024
+    return f"{bytes_value:.1f} TB"
+
+
+def format_eta(seconds):
+    """Formats seconds to human readable time string."""
+    if seconds is None or seconds < 0:
+        return "?"
+    if seconds < 60:
+        return f"{int(seconds)}s"
+    if seconds < 3600:
+        return f"{int(seconds // 60)}m {int(seconds % 60)}s"
+    return f"{int(seconds // 3600)}h {int((seconds % 3600) // 60)}m"
+
+
+async def safe_edit_message(query, text, reply_markup=None, parse_mode=None):
+    """Safely edit a Telegram message and ignore common transient failures."""
+
+    try:
+        await query.edit_message_text(
+            text,
+            reply_markup=reply_markup,
+            parse_mode=parse_mode,
+        )
+    except BadRequest as exc:
+        if "Message is not modified" not in str(exc):
+            raise
+    except (NetworkError, TimedOut) as exc:
+        logging.warning("Network error updating status message: %s", exc)
+
+
+async def send_long_message(bot, chat_id, text, header="", parse_mode="Markdown"):
+    """Split and send a long Telegram message in multiple chunks."""
+
+    max_length = 4000
+    parts = []
+    current = header
+
+    for line in text.split("\n"):
+        while len(line) > max_length:
+            split_at = max_length
+            for sep in [". ", "! ", "? ", ", ", " "]:
+                idx = line.rfind(sep, 0, max_length)
+                if idx > max_length // 2:
+                    split_at = idx + len(sep)
+                    break
+            if current.strip():
+                parts.append(current)
+                current = ""
+            parts.append(line[:split_at])
+            line = line[split_at:]
+
+        if len(current) + len(line) + 2 > max_length:
+            parts.append(current)
+            current = line + "\n"
+        else:
+            current += line + "\n"
+
+    if current.strip():
+        parts.append(current)
+
+    for part in parts:
+        try:
+            await bot.send_message(
+                chat_id=chat_id,
+                text=part,
+                parse_mode=parse_mode,
+                read_timeout=60,
+                write_timeout=60,
+            )
+        except BadRequest:
+            await bot.send_message(
+                chat_id=chat_id,
+                text=part,
+                read_timeout=60,
+                write_timeout=60,
+            )
